(chore) Consolidate appointment mocks to mono-repo root directory
