+++ __mocks__/appointments.mock.ts
@@ -264,42 +264,6 @@ export const mockAppointmentsData = {
   ],
 };
 
-export const mockServiceTypes = {
-  data: [
-    {
-      name: 'HIV Clinic',
-      uuid: '7bff050e-9a83-4fdb-9212-1a4c2cee349b',
-      duration: 15,
-    },
-    {
-      name: 'Drug Dispense',
-      uuid: '0e960ced-d35e-43de-b87f-bc08ef3b6ec3',
-      duration: 15,
-    },
-  ],
-};
-
-export const mockFrequency = {
-  data: [
-    { uuid: 'b127d100-bccb-4699-b30e-fe60a068ca46', display: 'Daily' },
-    { uuid: 'dbc32cf4-e11d-4006-b398-c71204cfc0d0', display: 'Weekly' },
-    { uuid: '6e263ecd-173e-4ee1-a9a5-df1f8d5b274b', display: 'Monthly' },
-  ],
-};
-
-export const mockProviders = {
-  data: [
-    {
-      uuid: 'f9badd80-ab76-11e2-9e96-0800200c9a66',
-      person: { uuid: '24252571-dd5a-11e6-9d9c-0242ac150002', display: 'Dr James Cook' },
-    },
-    {
-      uuid: '3191eddf-5cc5-4fa4-94ef-dfc25e8d33e4',
-      person: { uuid: '89af8ba6-2ec5-4d77-b3ed-7e9e02448e96', display: 'Dr Amstrong Neil' },
-    },
-  ],
-};
-
 export const mockMappedAppointmentsData = {
   data: [
     {
@@ -358,4 +322,55 @@ export const mockProvidersCount = {
 
 export const mockStartTime = {
   startTime: '2022-10-03T00:00:00.000+0000',
+};
+
+export const mockPatient = {
+  identifier: '100GEJ',
+  name: 'John Wilson',
+  uuid: '8673ee4f-e2ab-4077-ba55-4980f408773e',
+  gender: 'M',
+  age: 35,
+  birthdate: '1986-04-03T00:00:00.000+0000',
+  phoneNumber: '0700000000',
+};
+
+export const mockServices = [
+  { uuid: '176052c7-5fd4-4b33-89cc-7bae6848c65a', display: 'Clinical consultation' },
+  { uuid: 'd80ff12a-06a7-11ed-b939-0242ac120002', display: 'Triage' },
+];
+
+export const mockServiceTypes = {
+  data: [
+    {
+      name: 'HIV Clinic',
+      uuid: '7bff050e-9a83-4fdb-9212-1a4c2cee349b',
+      duration: 15,
+    },
+    {
+      name: 'Drug Dispense',
+      uuid: '0e960ced-d35e-43de-b87f-bc08ef3b6ec3',
+      duration: 15,
+    },
+  ],
+};
+
+export const mockFrequency = {
+  data: [
+    { uuid: 'b127d100-bccb-4699-b30e-fe60a068ca46', display: 'Daily' },
+    { uuid: 'dbc32cf4-e11d-4006-b398-c71204cfc0d0', display: 'Weekly' },
+    { uuid: '6e263ecd-173e-4ee1-a9a5-df1f8d5b274b', display: 'Monthly' },
+  ],
+};
+
+export const mockProviders = {
+  data: [
+    {
+      uuid: 'f9badd80-ab76-11e2-9e96-0800200c9a66',
+      person: { uuid: '24252571-dd5a-11e6-9d9c-0242ac150002', display: 'Dr James Cook' },
+    },
+    {
+      uuid: '3191eddf-5cc5-4fa4-94ef-dfc25e8d33e4',
+      person: { uuid: '89af8ba6-2ec5-4d77-b3ed-7e9e02448e96', display: 'Dr Amstrong Neil' },
+    },
+  ],
 };
