--- packages/esm-appointments-app/__mocks__/appointments.mock.ts
@@ -1,411 +1,0 @@
-// export const mockAppointmentsData = {
-//   data: [
-//     {
-//       uuid: '7cd38a6d-377e-491b-8284-b04cf8b8c6d8',
-//       appointmentNumber: '0000',
-//       patient: {
-//         identifier: '100GEJ',
-//         name: 'John Wilson',
-//         uuid: '8673ee4f-e2ab-4077-ba55-4980f408773e',
-//         gender: 'M',
-//         age: 35,
-//         birthdate: '1986-04-03T00:00:00.000+0000',
-//         phoneNumber: '0700000000',
-//       },
-//       service: {
-//         appointmentServiceId: 1,
-//         name: 'Outpatient',
-//         description: null,
-//         speciality: {},
-//         startTime: '',
-//         endTime: '',
-//         maxAppointmentsLimit: null,
-//         durationMins: null,
-//         location: {},
-//         uuid: 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90',
-//         initialAppointmentStatus: 'Scheduled',
-//         creatorName: null,
-//       },
-//       serviceType: {
-//         display: 'HIV Clinic',
-//         uuid: '53d58ff1-0c45-4e2e-9bd2-9cc826cb46e1',
-//         duration: 15,
-//       },
-//       provider: {
-//         uuid: 'f9badd80-ab76-11e2-9e96-0800200c9a66',
-//         person: { uuid: '24252571-dd5a-11e6-9d9c-0242ac150002', display: 'Dr James Cook' },
-//       },
-//       location: { name: 'HIV Clinic', uuid: '2131aff8-2e2a-480a-b7ab-4ac53250262b' },
-//       startDateTime: 1630326900000,
-//       endDateTime: 1630327200000,
-//       appointmentKind: 'WalkIn',
-//       status: 'Scheduled',
-//       comments: 'Walk in appointments',
-//       additionalInfo: null,
-//       providers: [],
-//       recurring: false,
-//     },
-//     {
-//       uuid: 'e10ce4e3-0e91-4b97-bc6c-9b5068e58428',
-//       appointmentNumber: '0000',
-//       patient: {
-//         identifier: '100GEJ',
-//         name: 'John Wilson',
-//         uuid: '8673ee4f-e2ab-4077-ba55-4980f408773e',
-//         gender: 'M',
-//         age: 35,
-//         birthdate: '1986-04-03T00:00:00.000+0000',
-//         phoneNumber: '0700000000',
-//       },
-//       service: {
-//         appointmentServiceId: 1,
-//         name: 'Outpatient',
-//         description: null,
-//         speciality: {},
-//         startTime: '',
-//         endTime: '',
-//         maxAppointmentsLimit: null,
-//         durationMins: null,
-//         location: {},
-//         uuid: 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90',
-//         initialAppointmentStatus: 'Scheduled',
-//         creatorName: null,
-//       },
-//       serviceType: {
-//         display: 'HIV Clinic',
-//         uuid: '53d58ff1-0c45-4e2e-9bd2-9cc826cb46e1',
-//         duration: 15,
-//       },
-//       provider: {
-//         uuid: 'f9badd80-ab76-11e2-9e96-0800200c9a66',
-//         person: { uuid: '24252571-dd5a-11e6-9d9c-0242ac150002', display: 'Dr James Cook' },
-//       },
-//       location: { name: 'HIV Clinic', uuid: '2131aff8-2e2a-480a-b7ab-4ac53250262b' },
-//       startDateTime: 1631278200000,
-//       endDateTime: 1631278560000,
-//       appointmentKind: 'WalkIn',
-//       status: 'Scheduled',
-//       comments: 'Some additional notes',
-//       additionalInfo: null,
-//       providers: [],
-//       recurring: false,
-//     },
-//     {
-//       uuid: 'cdb0676f-0805-4c3e-bfef-7757a005e892',
-//       appointmentNumber: '0000',
-//       patient: {
-//         identifier: '100GEJ',
-//         name: 'John Wilson',
-//         uuid: '8673ee4f-e2ab-4077-ba55-4980f408773e',
-//         gender: 'M',
-//         age: 35,
-//         birthdate: '1986-04-03T00:00:00.000+0000',
-//         phoneNumber: '0700000000',
-//       },
-//       service: {
-//         appointmentServiceId: 1,
-//         name: 'Outpatient',
-//         description: null,
-//         speciality: {},
-//         startTime: '',
-//         endTime: '',
-//         maxAppointmentsLimit: null,
-//         durationMins: null,
-//         location: {},
-//         uuid: 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90',
-//         initialAppointmentStatus: 'Scheduled',
-//         creatorName: null,
-//       },
-//       serviceType: {
-//         display: 'TB Clinic',
-//         uuid: '4a228e52-0bfe-11ed-861d-0242ac120002',
-//         duration: 15,
-//       },
-//       provider: null,
-//       location: { name: 'TB Clinic', uuid: '2131aff8-2e2a-480a-b7ab-4ac53250262b' },
-//       startDateTime: 1631537400000,
-//       endDateTime: 1631537760000,
-//       appointmentKind: 'WalkIn',
-//       status: 'Scheduled',
-//       comments: 'Some additional notes',
-//       additionalInfo: null,
-//       providers: [],
-//       recurring: false,
-//     },
-//     {
-//       uuid: '66565d8b-4849-4b7c-966a-554d6073f80c',
-//       appointmentNumber: '0000',
-//       patient: {
-//         identifier: '100GEJ',
-//         name: 'John Wilson',
-//         uuid: '8673ee4f-e2ab-4077-ba55-4980f408773e',
-//         gender: 'M',
-//         age: 35,
-//         birthdate: '1986-04-03T00:00:00.000+0000',
-//         phoneNumber: '0700000000',
-//       },
-//       service: {
-//         appointmentServiceId: 1,
-//         name: 'Outpatient',
-//         description: null,
-//         speciality: {},
-//         startTime: '',
-//         endTime: '',
-//         maxAppointmentsLimit: null,
-//         durationMins: null,
-//         location: {},
-//         uuid: 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90',
-//         initialAppointmentStatus: 'Scheduled',
-//         creatorName: null,
-//       },
-//       serviceType: {
-//         display: 'HIV Clinic',
-//         uuid: '53d58ff1-0c45-4e2e-9bd2-9cc826cb46e1',
-//         duration: 15,
-//       },
-//       provider: null,
-//       location: { name: 'HIV Clinic', uuid: '2131aff8-2e2a-480a-b7ab-4ac53250262b' },
-//       startDateTime: 1631605800000,
-//       endDateTime: 1631606100000,
-//       appointmentKind: 'WalkIn',
-//       status: 'Scheduled',
-//       comments: 'Some additional notes',
-//       additionalInfo: null,
-//       providers: [],
-//       recurring: false,
-//     },
-//     {
-//       uuid: '45dcc19d-dd14-4a07-95c6-afa264972a34',
-//       appointmentNumber: '0000',
-//       patient: {
-//         identifier: '100GEJ',
-//         name: 'John Wilson',
-//         uuid: '8673ee4f-e2ab-4077-ba55-4980f408773e',
-//         gender: 'M',
-//         age: 35,
-//         birthdate: '1986-04-03T00:00:00.000+0000',
-//         phoneNumber: '0700000000',
-//       },
-//       service: {
-//         appointmentServiceId: 1,
-//         name: 'Outpatient',
-//         description: null,
-//         speciality: {},
-//         startTime: '',
-//         endTime: '',
-//         maxAppointmentsLimit: null,
-//         durationMins: null,
-//         location: {},
-//         uuid: 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90',
-//         initialAppointmentStatus: 'Scheduled',
-//         creatorName: null,
-//       },
-//       serviceType: {
-//         display: 'TB Clinic',
-//         uuid: '4a228e52-0bfe-11ed-861d-0242ac120002',
-//         duration: 15,
-//       },
-//       provider: {
-//         uuid: 'f9badd80-ab76-11e2-9e96-0800200c9a66',
-//         person: { uuid: '24252571-dd5a-11e6-9d9c-0242ac150002', display: 'Dr James Cook' },
-//       },
-//       location: { name: 'TB Clinic', uuid: '2131aff8-2e2a-480a-b7ab-4ac53250262b' },
-//       startDateTime: 1631623800000,
-//       endDateTime: 1631624160000,
-//       appointmentKind: 'WalkIn',
-//       status: 'Scheduled',
-//       comments: 'Some additional notes',
-//       additionalInfo: null,
-//       providers: [],
-//       recurring: false,
-//     },
-//     {
-//       uuid: 'fa4657ad-db46-487d-8e2e-a3858c906ae6',
-//       appointmentNumber: '0000',
-//       patient: {
-//         identifier: '100GEJ',
-//         name: 'John Wilson',
-//         uuid: '8673ee4f-e2ab-4077-ba55-4980f408773e',
-//         gender: 'M',
-//         age: 35,
-//         birthdate: '1986-04-03T00:00:00.000+0000',
-//         phoneNumber: '0700000000',
-//       },
-//       service: {
-//         appointmentServiceId: 1,
-//         name: 'Outpatient',
-//         description: null,
-//         speciality: {},
-//         startTime: '',
-//         endTime: '',
-//         maxAppointmentsLimit: null,
-//         durationMins: null,
-//         location: {},
-//         uuid: 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90',
-//         initialAppointmentStatus: 'Scheduled',
-//         creatorName: null,
-//       },
-//       serviceType: {
-//         display: 'TB Clinic',
-//         uuid: '4a228e52-0bfe-11ed-861d-0242ac120002',
-//         duration: 15,
-//       },
-//       provider: null,
-//       location: { name: 'TB Clinic', uuid: '2131aff8-2e2a-480a-b7ab-4ac53250262b' },
-//       startDateTime: 1631712720000,
-//       endDateTime: 1631713080000,
-//       appointmentKind: 'WalkIn',
-//       status: 'Scheduled',
-//       comments: 'Some value',
-//       additionalInfo: null,
-//       providers: [],
-//       recurring: false,
-//     },
-//   ],
-// };
-
-// export const mockAppointments = {
-//   pastAppointments: [
-//     {
-//       uuid: 'a4662406-2bfe-49be-8465-6dbdfe9b4d5d',
-//       appointmentNumber: '0000',
-//       patient: {
-//         identifier: '1003R8',
-//         name: 'Brian Evans',
-//         uuid: '8f75dbf1-833d-498a-bc99-3a24cad5cee9',
-//       },
-//       service: {
-//         appointmentServiceId: 1,
-//         name: 'HIV return visit',
-//         description: null,
-//         speciality: {},
-//         startTime: '',
-//         endTime: '',
-//         maxAppointmentsLimit: null,
-//         durationMins: null,
-//         location: {},
-//         uuid: 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90',
-//         color: '#006400',
-//         initialAppointmentStatus: 'Scheduled',
-//         creatorName: null,
-//       },
-//       serviceType: null,
-//       provider: null,
-//       location: {
-//         name: 'Outpatient Clinic',
-//         uuid: 'f76c0c8e-2c3a-443c-b26d-96a9f3847764',
-//       },
-//       startDateTime: '2021-10-28T11:14:17.064Z',
-//       endDateTime: 1649935800000,
-//       appointmentKind: 'Scheduled',
-//       status: 'Scheduled',
-//       comments: null,
-//       additionalInfo: null,
-//       teleconsultation: null,
-//       providers: [],
-//       voided: false,
-//       extensions: {
-//         patientEmailDefined: false,
-//       },
-//       teleconsultationLink: null,
-//       recurring: false.valueOf,
-//     },
-//   ],
-//   upcomingAppointments: [
-//     {
-//       uuid: 'a4662406-2bfe-49be-8465-6dbdfe9b4d5d',
-//       appointmentNumber: '0000',
-//       patient: {
-//         identifier: '1003R8',
-//         name: 'Brian Evans',
-//         uuid: '8f75dbf1-833d-498a-bc99-3a24cad5cee9',
-//       },
-//       service: {
-//         appointmentServiceId: 1,
-//         name: 'Clinical observation',
-//         description: null,
-//         speciality: {},
-//         startTime: '',
-//         endTime: '',
-//         maxAppointmentsLimit: null,
-//         durationMins: null,
-//         location: {},
-//         uuid: 'e2ec9cf0-ec38-4d2b-af6c-59c82fa30b90',
-//         color: '#006400',
-//         initialAppointmentStatus: 'Scheduled',
-//         creatorName: null,
-//       },
-//       serviceType: null,
-//       provider: null,
-//       location: {
-//         name: 'Outpatient Clinic',
-//         uuid: 'f76c0c8e-2c3a-443c-b26d-96a9f3847764',
-//       },
-//       startDateTime: 1649934000000,
-//       endDateTime: 1649935800000,
-//       appointmentKind: 'Scheduled',
-//       status: 'Scheduled',
-//       comments: null,
-//       additionalInfo: null,
-//       teleconsultation: null,
-//       providers: [],
-//       voided: false,
-//       extensions: {
-//         patientEmailDefined: false,
-//       },
-//       teleconsultationLink: null,
-//       recurring: false.valueOf,
-//     },
-//   ],
-// };
-
-export const mockPatient = {
-  identifier: '100GEJ',
-  name: 'John Wilson',
-  uuid: '8673ee4f-e2ab-4077-ba55-4980f408773e',
-  gender: 'M',
-  age: 35,
-  birthdate: '1986-04-03T00:00:00.000+0000',
-  phoneNumber: '0700000000',
-};
-
-export const mockServices = [
-  { uuid: '176052c7-5fd4-4b33-89cc-7bae6848c65a', display: 'Clinical consultation' },
-  { uuid: 'd80ff12a-06a7-11ed-b939-0242ac120002', display: 'Triage' },
-];
-
-export const mockServiceTypes = {
-  data: [
-    {
-      name: 'HIV Clinic',
-      uuid: '7bff050e-9a83-4fdb-9212-1a4c2cee349b',
-      duration: 15,
-    },
-    {
-      name: 'Drug Dispense',
-      uuid: '0e960ced-d35e-43de-b87f-bc08ef3b6ec3',
-      duration: 15,
-    },
-  ],
-};
-
-export const mockFrequency = {
-  data: [
-    { uuid: 'b127d100-bccb-4699-b30e-fe60a068ca46', display: 'Daily' },
-    { uuid: 'dbc32cf4-e11d-4006-b398-c71204cfc0d0', display: 'Weekly' },
-    { uuid: '6e263ecd-173e-4ee1-a9a5-df1f8d5b274b', display: 'Monthly' },
-  ],
-};
-
-export const mockProviders = {
-  data: [
-    {
-      uuid: 'f9badd80-ab76-11e2-9e96-0800200c9a66',
-      person: { uuid: '24252571-dd5a-11e6-9d9c-0242ac150002', display: 'Dr James Cook' },
-    },
-    {
-      uuid: '3191eddf-5cc5-4fa4-94ef-dfc25e8d33e4',
-      person: { uuid: '89af8ba6-2ec5-4d77-b3ed-7e9e02448e96', display: 'Dr Amstrong Neil' },
-    },
-  ],
-};
